use threads

+++ script.py
@@ -8,7 +8,7 @@ NUM_REPLICATES = 100_000
 
 print(
     bootstrap(
-        data=(np.arange(N),),
+        data=(np.arange(N, dtype=np.float64),),
         statistic=partial(np.var, ddof=1),
         n_resamples=NUM_REPLICATES,
     ).standard_error
